requirements.txt
apify-client
langchain
langchain-openai
langchain-groq
langchain-ollama
langchain-nvidia-ai-endpoints
langchain-google-genai
langchain-community
python-docx
pandas
setuptools
matplotlib
matplotlib-inline
seaborn
python-dotenv
numpy
scikit-learn
hyperopt
wordcloud
unstructured
openpyxl
apify-client
serpapi
google-generativeai
uvicorn
streamlit 
httpx
sqlalchemy
sqlalchemy-utils
fastapi
passlib[bcrypt]
python-jose[cryptography]
langchain-core
pydantic
pydantic-settings
requests